backend/requirements.txt
openai
python-dotenv
uvicorn
fastapi
pydantic
